+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## Unpublished
 ### Changed
 - show full formatted news in news overview
+- update news form buttons to button dropdown
 
 ## v0.7.1 - 2024-03-21
 ### Added